+++ ruoyi-system/src/main/java/com/ruoyi/system/service/SysReserveService.java
@@ -22,7 +22,7 @@ public interface SysReserveService {
 
     int updateReserveStatus(BusinessReserve businessReserve);
 
-    int deleteReserveById(Long id);
+    int deleteReserveById(Integer id);
 
     BusinessReserve selectReserveById(Integer id);
 }
